+++ app.js
@@ -3,7 +3,7 @@ const express = require('express')
 const logger = require('morgan')
 const cors = require('cors')
 
-const contactsRouter = require('./routes/api/contacts')
+const contactsRouter = require('./routes/api/contactsRoutes')
 
 const app = express()
